updating 'absence' view structure, adding attandeces for 'ACIERIE' projet section, adding services management (create, delete)
