+++ resources/css/app.css
@@ -831,3 +831,7 @@ button:active{
   display: inline-block;
   width: 30% !important;
 }
+
+.dt-container .top{
+  padding: 10px;
+}
